Change #contents to #in
After actual use I realize `#contents` is used a LOT.

Typing so much is impractical and verbose, not to mention the
oversized line-lengths.

`#in` is the new-and-improved version. Much shorter and manages to
keep the meaning clear. (There weren't really that many options
to replace `#contents`. I checked a thesaurus.)