Add support for emitting raw strings from render arrays
Sometimes you want to weight a plain string or preprocess it with
callbacks. If you wrap a plain string in a render array it will
always be enclosed in tags of some sort.

While you could write a callback that returns the #in value, or
set the #tag to something benign and inline like "span", this is
extra work compared to simply adding a #raw element which is
returned regardless of the state of the array.